Extract code to filter watched artifacts
