+++ pkg/skaffold/runner/dev_test.go
@@ -23,7 +23,6 @@ import (
 	"testing"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/config"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/test"
@@ -191,55 +190,5 @@ func TestBuildAndDeployAllArtifacts(t *testing.T) {
 	}
 	if len(deployer.deployed) != 2 {
 		t.Errorf("Expected 2 artifacts to be deployed. Got %d", len(deployer.deployed))
-	}
-}
-
-func TestShouldWatch(t *testing.T) {
-	var tests = []struct {
-		description   string
-		watch         []string
-		expectedMatch bool
-	}{
-		{
-			description:   "match all",
-			watch:         nil,
-			expectedMatch: true,
-		},
-		{
-			description:   "match full name",
-			watch:         []string{"domain/image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "match partial name",
-			watch:         []string{"image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "match any",
-			watch:         []string{"other", "image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "no match",
-			watch:         []string{"other"},
-			expectedMatch: false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.description, func(t *testing.T) {
-			runner := &SkaffoldRunner{
-				opts: &config.SkaffoldOptions{
-					Watch: test.watch,
-				},
-			}
-
-			match := runner.shouldWatch(&latest.Artifact{
-				ImageName: "domain/image",
-			})
-
-			testutil.CheckDeepEqual(t, test.expectedMatch, match)
-		})
 	}
 }
