+++ pkg/skaffold/runner/filters.go
@@ -1,0 +1,37 @@
+/*
+Copyright 2018 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package runner
+
+import (
+	"strings"
+
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
+)
+
+func (r *SkaffoldRunner) shouldWatch(artifact *latest.Artifact) bool {
+	if len(r.opts.Watch) == 0 {
+		return true
+	}
+
+	for _, watchExpression := range r.opts.Watch {
+		if strings.Contains(artifact.ImageName, watchExpression) {
+			return true
+		}
+	}
+
+	return false
+}
